// src/store/type.d.ts
export interface ForwardStore {
  // 合并转发的状态
  selectionMode: boolean;
  selectedMessages: MessageItem[];

  // 操作方法
  startSelectionMode: () => void;
  cancelSelectionMode: () => void;
  toggleMessage: (message: MessageItem) => void;
  selectMessages: (messages: MessageItem[]) => void;
  clearSelectedMessages: () => void;
  batchForwardMessages: ({
    isGroup,
    peerNickname,
    selfNickname,
  }: {
    isGroup: boolean;
    peerNickname?: string;
    selfNickname?: string;
  }) => Promise<void>;
}